05/07/2023
Se cambia el float por int

+++ Tipo_de_datos.py
@@ -1,3 +1,3 @@
-a= float(input("ingresda numero a: " ))
+a= int(input("ingresda numero a: " ))
 b= int(input("ingresda numero b: " ))
 print(a+b)
